Schemes generation (#188)
* Return GeneratedProject from ProjectGenerator

* Add Launch, Profile, Analyze and Archive actions

* Add BuildAction generation

* Generate TestsAction

* Add some tests

* Add Cucumber test

* Remove byebug line

* Update CHANGELOG

* Fix test

+++ features/step_definitions/shared/workspace.rb
@@ -5,6 +5,6 @@ And(/I have have a working directory/) do
   @dir = Dir.mktmpdir
 end
 
-Then(/I delete the working directory/) do
-  FileUtils.rm_r(@dir)
+After do
+  FileUtils.rm_r(@dir) unless @dir.nil?
 end
